Refactor file removal into RemoveFileController
The responsibilities of the ReviewFileController were getting muddled
with reviewing and deleting an uploaded file. Moving removal to a
separate controller is clearer and is more idiomatic.
