+++ app/components/question/file_remove_component/view.html.erb
@@ -1,4 +1,4 @@
-<h1 class="govuk-heading-l"><%= t('forms.review_file.confirmation.title') %></h1>
+<h1 class="govuk-heading-l"><%= t('forms.remove_file.show.title') %></h1>
 
 <p class="govuk-inset-text">
   <%= question.original_filename %>
@@ -13,7 +13,7 @@
 
       <%= form.govuk_collection_radio_buttons :remove,
             @remove_input.values, ->(option) { option }, ->(option) { t('helpers.label.remove_input.options.' + "#{option}") },
-            legend: { text: t('forms.review_file.confirmation.radios_legend')},
+            legend: { text: t('forms.remove_file.show.radios_legend')},
             inline: true  %>
 
       <%= form.govuk_submit  %>
