+++ config/locales/en.yml
@@ -159,10 +159,11 @@ en:
         phone_number_heading: Remove phone number
         phone_number_radios_legend: Are you sure you want to remove this phone number?
         short_answer_legend: Are you sure you want to remove ‘%{answer_text}’?
-    review_file:
-      confirmation:
+    remove_file:
+      show:
         radios_legend: Are you sure you want to remove your file?
         title: Remove your file
+    review_file:
       show:
         check_file: Check your uploaded file
         remove: Remove
