+++ spec/requests/forms/review_file_controller_spec.rb
@@ -108,118 +108,6 @@ RSpec.describe Forms::ReviewFileController, type: :request do
     end
   end
 
-  describe "#delete" do
-    let(:mock_s3_client) { Aws::S3::Client.new(stub_responses: true) }
-    let(:remove) { "yes" }
-
-    before do
-      allow(Aws::S3::Client).to receive(:new).and_return(mock_s3_client)
-      allow(mock_s3_client).to receive(:delete_object)
-      delete remove_file_path(mode:, form_id: form_data.id, form_slug: form_data.form_slug, page_slug:, changing_existing_answer:, remove_input: { remove: })
-    end
-
-    context "when the question is a file upload question" do
-      let(:page_slug) { file_upload_step.id.to_s }
-
-      context "when the input object validation fails" do
-        let(:remove) { "invalid" }
-
-        it "renders the confirmation page with 422 status" do
-          expect(response).to render_template("forms/review_file/confirmation")
-          expect(response).to have_http_status :unprocessable_entity
-        end
-
-        it "displays an error" do
-          rendered = Capybara.string(response.body)
-          expect(rendered).to have_css(".govuk-error-summary")
-        end
-
-        it "displays a back link to the review file page" do
-          expect(response.body).to include(review_file_path(form_data.id, form_data.form_slug, page_slug, changing_existing_answer:))
-        end
-      end
-
-      context "when the user has confirmed they want to remove their file" do
-        context "when a file has been uploaded" do
-          it "deletes the file from S3" do
-            expect(mock_s3_client).to have_received(:delete_object)
-          end
-
-          it "removes the answer from the session" do
-            expect(store[:answers][form_data.id.to_s]).not_to have_key page_slug
-          end
-
-          it "redirects to the show page route" do
-            expect(response).to redirect_to form_page_path(form_data.id, form_data.form_slug, page_slug)
-          end
-
-          it "displays a success banner" do
-            expect(flash[:success]).to eq(I18n.t("banner.success.file_removed"))
-          end
-
-          context "when changing an existing answer" do
-            let(:changing_existing_answer) { true }
-
-            it "redirects to the change answer route" do
-              expect(response).to redirect_to form_change_answer_path(form_data.id, form_data.form_slug, page_slug)
-            end
-          end
-        end
-
-        context "when a file has not been uploaded" do
-          let(:uploaded_file_key) { nil }
-
-          it "does not remove the answer from the session" do
-            expect(store[:answers][form_data.id.to_s]).to have_key page_slug
-          end
-
-          it "redirects to the show page route" do
-            expect(response).to redirect_to form_page_path(form_data.id, form_data.form_slug, page_slug)
-          end
-        end
-      end
-
-      context "when the user has not confirmed they want to remove their file" do
-        let(:remove) { "no" }
-
-        it "does not delete the file from S3" do
-          expect(mock_s3_client).not_to have_received(:delete_object)
-        end
-
-        it "does not remove the answer from the session" do
-          expect(store[:answers][form_data.id.to_s]).to have_key page_slug
-        end
-
-        it "redirects to the review file page route" do
-          expect(response).to redirect_to review_file_path(form_data.id, form_data.form_slug, page_slug, changing_existing_answer:)
-        end
-      end
-    end
-
-    context "when the question isn't a file upload question" do
-      let(:page_slug) { text_question_step.id.to_s }
-      let(:store) do
-        {
-          answers: {
-            form_data.id.to_s => {
-              text_question_step.id.to_s => {
-                "text" => "foo",
-              },
-            },
-          },
-        }
-      end
-
-      it "does not remove the answer from the session" do
-        expect(store[:answers][form_data.id.to_s]).to have_key page_slug
-      end
-
-      it "redirects to the show page route" do
-        expect(response).to redirect_to form_page_path(form_data.id, form_data.form_slug, page_slug)
-      end
-    end
-  end
-
   describe "#continue" do
     before do
       post review_file_continue_path(mode:, form_id: form_data.id, form_slug: form_data.form_slug, page_slug:, changing_existing_answer:)
@@ -247,56 +135,6 @@ RSpec.describe Forms::ReviewFileController, type: :request do
 
         it "redirects to the show page route" do
           expect(response).to redirect_to form_page_path(form_data.id, form_data.form_slug, page_slug)
-        end
-      end
-    end
-
-    context "when the question isn't a file upload question" do
-      let(:page_slug) { text_question_step.id }
-
-      it "redirects to the show page route" do
-        expect(response).to redirect_to form_page_path(form_data.id, form_data.form_slug, page_slug)
-      end
-    end
-  end
-
-  describe "#confirmation" do
-    before do
-      get remove_file_confirmation_path(mode:, form_id: form_data.id, form_slug: form_data.form_slug, page_slug:, changing_existing_answer:)
-    end
-
-    context "when the question is a file upload question" do
-      let(:page_slug) { file_upload_step.id }
-
-      context "when a file has been uploaded" do
-        it "renders the remove file confirmation template" do
-          expect(response).to render_template("forms/review_file/confirmation")
-        end
-
-        it "displays the uploaded filename" do
-          expect(response.body).to include(uploaded_filename)
-        end
-
-        it "displays a back link to the review file page" do
-          expect(response.body).to include(review_file_path(form_data.id, form_data.form_slug, page_slug, changing_existing_answer:))
-        end
-      end
-
-      context "when a file has not been uploaded" do
-        let(:store) { {} }
-
-        it "redirects to the show page route" do
-          expect(response).to redirect_to form_page_path(form_data.id, form_data.form_slug, page_slug)
-        end
-      end
-
-      context "when changing an existing answer" do
-        let(:changing_existing_answer) { true }
-
-        it "includes the changing_existing_answer query parameter for the confirmation URL" do
-          rendered = Capybara.string(response.body)
-          expected_url = remove_file_confirmation_path(mode:, form_id: form_data.id, form_slug: form_data.form_slug, page_slug:, changing_existing_answer:)
-          expect(rendered).to have_css("form[action='#{expected_url}'][method='post']")
         end
       end
     end
